Indexer: Reduce store time to less than 50%
Previously the indexer was bottle-necked by the speed of the database.
One reason for that was the index on the words.

By adding the index after we inserted all rows, we could decrease the
time spent in PutAllWords from 56% of the total CPU time to just 24%.

This change introduces a new function to the documentstore interface,
which is not perfect, but the performance definitely speaks for itself
here. ^^